+++ CHANGELOG.md
@@ -6,9 +6,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2026-02-17
+## [0.1.1] - 2026-02-25
 
 ### Added
+- new error type `ListError`
+
+### Changed
+- to_vec() improved
+- get() and get_mut() improved
+
+### Fixed
+- ci.yaml
+- this file
+- find_if() - unnecessary restriction removed 
+
+## [0.1.0] - 2026-02-25
+
+### Added
+- `SingleLinkedList` lite implementation
 
 ### Changed